fixed onboarding bug
