+++ src/components/quizlet/infoScreen.tsx
@@ -14,11 +14,6 @@ export interface InfoScreenProps {
   handleFormSubmit: Function;
 }
 
-// alert for snackbar
-function Alert(props:any) {
-  return <MuiAlert elevation={6} variant="filled" {...props} />;
-}
-
 const useStyles = makeStyles((theme) => ({
   inputContainer: {
     width: "80%",
@@ -37,72 +32,10 @@ const useStyles = makeStyles((theme) => ({
 
 const InfoScreen: React.FC<InfoScreenProps> = (props) => {
   const classes = useStyles();
-  
-  // initial state for snackbar
-  const [open, setOpen] = React.useState(false);
 
   const { handleFormSubmit, handleChange, handleNextScreen } = props;
   const { data: formData, display: formDisplay, focus } = props.formState;
   useEnterKeyPress(() => handleFormSubmit());
-
-  // snackbar link 
-  // https://material-ui.com/components/snackbars/
-   // snackbar onClick
-  const handleClick = () => {
-    setOpen(true);
-  };
-
-  // snackbar close
-  const handleClose = (event: any, reason: string) => {
-    if (reason === 'clickaway') {
-      return;
-    }
-    setOpen(false);
-  };
-
-  // API connection
-  // const saveFormData = async () => {
-  //   const response = await fetch('https://dev.intelligems.io/track', {
-  //     method: 'POST',
-  //     body: JSON.stringify(formData)
-  //   });
-  //   if (response.status !== 200) {
-  //     throw new Error(`Request failed: ${response.status}`); 
-  //   } else{
-  //     console.log(response.status);
-  //   }
-  // }
-  // // data save
-  // const onSubmit = async (event:any) => {
-  //   event.preventDefault(); // Prevent default submission
-  //   try {
-  //     await saveFormData();
-  //     console.log('Your registration was successfully submitted!');
-  //     console.log(formData)
-  //     setValues(formData);
-  //   } catch (e) {
-  //     alert(`Registration failed! ${e.message}`);
-  //   }
-  // }
-
-// one of my previous fetch versions, unsure how efficient this is since it is not in an await or using async
-  // fetch('https://dev.intelligems.io/track', {
-  //   method: 'POST',
-  //   body: JSON.stringify(formData),
-  //   headers: {
-  //       'Content-Type': 'application/json'
-  //   }
-  // }).then(response => {
-  //   if (response.status !== 200) {
-  //     console.log('Somthing happened wrong');
-  //     } else {
-  //       // console.log(response);
-  //       // console.log(response.status);
-  //       // console.log(formData)
-  //      return response;
-  //     }
-  // }).catch(err => err);
-
 
   return (
     <div>
@@ -119,7 +52,6 @@ const InfoScreen: React.FC<InfoScreenProps> = (props) => {
           onChange={handleChange}
           show={formDisplay.name}
           focus={focus}
-          
         />
         <InputRow
           question="2. What is your email address?"
@@ -146,18 +78,10 @@ const InfoScreen: React.FC<InfoScreenProps> = (props) => {
             size="large"
             color="secondary"
             type="submit"
-            onClick = {()=>{ handleClick()}}
-            // add inside onclick to loop back around to welcome screen
-            // handleNextScreen()
           >
             Continue
           </Button>
           or press <strong>ENTER</strong>
-            <Snackbar open={open} autoHideDuration={3000} onClose={handleClose}>
-            <Alert onClose={handleClose} severity="success">
-            Thanks! Your information has been saved!
-            </Alert>
-          </Snackbar>
         </div>
       </form>
     </div>
