+++ src/utils/numberFormats.tsx
@@ -49,3 +49,27 @@ export function PercentageFormat(props: CustomFormatProps) {
     />
   );
 }
+
+export function RegularFormat(props: CustomFormatProps) {
+  const { inputRef, onChange, ...other } = props;
+
+  return (
+    <NumberFormat
+      {...other}
+      getInputRef={inputRef}
+      onValueChange={(values) => {
+        onChange({
+          target: {
+            name: props.name,
+            value: values.value,
+          },
+        });
+      }}
+      thousandSeparator
+      isNumericString
+    />
+  );
+}
+
+
+
